Work in progress.  Need to split the Forms from the underlying logic so the display is a separate thing from the Object(s) being displayed.

+++ AddressBook.cs
@@ -16,7 +16,7 @@ using Newtonsoft.Json.Converters;
 
 namespace TDX_Extended
 {
-    public partial class AddressBook : Form
+    internal partial class AddressBook : Form
     {
 
         public enum Selection
@@ -36,61 +36,67 @@ namespace TDX_Extended
         public string sandboxUserAddress { get; set; }
         public string sandboxUserUsername { get; set; }
         public string sandboxUserPwd { get; set; }
-        //public Selection selection { get; set; }
+        public Selection selection { get; set; }
+
+        private TDXE globalSettings;
 
 
-        public AddressBook(Selection defaultChoice)
+        public AddressBook(TDXE settings)
         {
 
             InitializeComponent();
 
-            switch (defaultChoice)
+            globalSettings = settings;
+
+            switch ((AddressBook.Selection) settings.addressBookSelection)
             {
                 case Selection.prodServer: prodAdminBtn.Checked = true; break;
                 case Selection.prodUser: prodUserBtn.Checked = true; break;
                 case Selection.sandboxServer: sandboxAdminBtn.Checked = true; break;
                 default: sandboxUserBtn.Checked = true; break;
             }
- 
 
-            var builder = new ConfigurationBuilder()
-                .SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("secrets.json", optional: false, reloadOnChange: true)
-                .AddUserSecrets<AddressBook>().Build();
+            var secretsID = "48024d93-9b5e-40b8-99a1-e75d6cf33c61";
+            var secretsPath = Microsoft.Extensions.Configuration.UserSecrets.PathHelper.GetSecretsPathFromSecretsId(secretsID);
 
-            sandboxServerAddress = builder.GetSection("sandboxServerAddress").Value;
-            sandboxServerBEID = builder.GetSection("sandboxServerBEID").Value;
-            sandboxServerKey = builder.GetSection("sandboxServerKey").Value;
-            prodServerAddress = builder.GetSection("prodServerAddress").Value;
-            prodServerBEID = builder.GetSection("prodServerBEID").Value;
-            prodServerKey = builder.GetSection("prodServerKey").Value;
-            sandboxUserAddress = builder.GetSection("sandboxUserAddress").Value;
-            sandboxUserUsername = builder.GetSection("sandboxUserUsername").Value;
-            sandboxUserPwd = builder.GetSection("sandboxUserPwd").Value;
-            prodUserAddress = builder.GetSection("prodUserAddress").Value;
-            prodUserUsername = builder.GetSection("prodUserUsername").Value;
-            prodUserPwd = builder.GetSection("prodUserPwd").Value;
+            if (File.Exists(secretsPath)) {
+                var secretsJson = File.ReadAllText(secretsPath);
+                dynamic secrets = JsonConvert.DeserializeObject<ExpandoObject>(secretsJson, new ExpandoObjectConverter());
+
+                sandboxServerAddress = secrets.sandboxServerAddress;
+                sandboxServerBEID = secrets.sandboxServerBEID;
+                sandboxServerKey = secrets.sandboxServerKey;
+                prodServerAddress = secrets.prodServerAddress;
+                prodServerBEID = secrets.prodServerBEID;
+                prodServerKey = secrets.prodServerKey;
+                sandboxUserAddress = secrets.sandboxUserAddress;
+                sandboxUserUsername = secrets.sandboxUserUsername;
+                sandboxUserPwd = secrets.sandboxUserPwd;
+                prodUserAddress = secrets.prodUserAddress;
+                prodUserUsername = secrets.prodUserUsername;
+                prodUserPwd = secrets.prodUserPwd;
 
 
-            sandboxServerTxtBox.Text = sandboxServerAddress;
-            sandboxBeidTxtBox.Text = sandboxServerBEID;
-            sandboxKeyTxtBox.Text = sandboxServerKey;
-            sandboxUserServerTxtBox.Text = sandboxUserAddress;
-            sandboxUserUsernameTxtBox.Text = sandboxUserUsername;
-            sandboxUserPwdTxtBox.Text = sandboxUserPwd;
+                sandboxServerTxtBox.Text = sandboxServerAddress;
+                sandboxBeidTxtBox.Text = sandboxServerBEID;
+                sandboxKeyTxtBox.Text = sandboxServerKey;
+                sandboxUserServerTxtBox.Text = sandboxUserAddress;
+                sandboxUserUsernameTxtBox.Text = sandboxUserUsername;
+                sandboxUserPwdTxtBox.Text = sandboxUserPwd;
 
-            prodServerTxtBox.Text = prodServerAddress;
-            prodBeidTxtBox.Text = prodServerBEID;
-            prodKeyTxtBox.Text = prodServerKey;
-            prodUserServerTxtBox.Text = prodUserAddress;
-            prodUserUsernameTxtBox.Text = prodUserUsername;
-            prodUserPwdTxtBox.Text = prodUserPwd;
+                prodServerTxtBox.Text = prodServerAddress;
+                prodBeidTxtBox.Text = prodServerBEID;
+                prodKeyTxtBox.Text = prodServerKey;
+                prodUserServerTxtBox.Text = prodUserAddress;
+                prodUserUsernameTxtBox.Text = prodUserUsername;
+                prodUserPwdTxtBox.Text = prodUserPwd;
+            }
 
         }
 
         private void cancelBtn_Click(object sender, EventArgs e)
         {
-            this.Close();
+            this.Visible = false;
         }
 
         private void prodAdminBtn_CheckedChanged(object sender, EventArgs e)
@@ -141,25 +147,64 @@ namespace TDX_Extended
         {
             var secretsID = "48024d93-9b5e-40b8-99a1-e75d6cf33c61";
             var secretsPath = Microsoft.Extensions.Configuration.UserSecrets.PathHelper.GetSecretsPathFromSecretsId(secretsID);
-            var secretsJson = File.ReadAllText(secretsPath);
-            dynamic secrets = JsonConvert.DeserializeObject<ExpandoObject>(secretsJson, new ExpandoObjectConverter());
+            if (File.Exists(secretsPath))
+            {
+                var secretsJson = File.ReadAllText(secretsPath);
+                dynamic secrets = JsonConvert.DeserializeObject<ExpandoObject>(secretsJson, new ExpandoObjectConverter());
 
-            secrets.sandboxServerAddress = sandboxServerTxtBox.Text;
-            secrets.sandboxServerBEID = sandboxBeidTxtBox.Text;
-            secrets.sandboxServerKey = sandboxKeyTxtBox.Text;
-            secrets.sandboxUserAddress = sandboxUserServerTxtBox.Text;
-            secrets.sandboxUserUsername = sandboxUserUsernameTxtBox.Text;
-            secrets.sandboxUserPwd = sandboxUserPwdTxtBox.Text;
+                secrets.sandboxServerAddress = sandboxServerTxtBox.Text;
+                secrets.sandboxServerBEID = sandboxBeidTxtBox.Text;
+                secrets.sandboxServerKey = sandboxKeyTxtBox.Text;
+                secrets.sandboxUserAddress = sandboxUserServerTxtBox.Text;
+                secrets.sandboxUserUsername = sandboxUserUsernameTxtBox.Text;
+                secrets.sandboxUserPwd = sandboxUserPwdTxtBox.Text;
 
-            secrets.prodServerAddress = prodServerTxtBox.Text;
-            secrets.prodServerBEID = prodBeidTxtBox.Text;
-            secrets.prodServerKey = prodKeyTxtBox.Text;
-            secrets.prodUserAddress = prodUserServerTxtBox.Text;
-            secrets.prodUserUsername = prodUserUsernameTxtBox.Text;
-            secrets.prodUserPwd = prodUserPwdTxtBox.Text;
+                secrets.prodServerAddress = prodServerTxtBox.Text;
+                secrets.prodServerBEID = prodBeidTxtBox.Text;
+                secrets.prodServerKey = prodKeyTxtBox.Text;
+                secrets.prodUserAddress = prodUserServerTxtBox.Text;
+                secrets.prodUserUsername = prodUserUsernameTxtBox.Text;
+                secrets.prodUserPwd = prodUserPwdTxtBox.Text;
 
-            var updatedSecretsJson = JsonConvert.SerializeObject(secrets, Formatting.Indented);
-            File.WriteAllText(secretsPath, updatedSecretsJson);
+                var updatedSecretsJson = JsonConvert.SerializeObject(secrets, Formatting.Indented);
+                File.WriteAllText(secretsPath, updatedSecretsJson);
+            }
+            else
+            {
+                dynamic secrets = new ExpandoObject();
+
+                secrets.sandboxServerAddress = sandboxServerTxtBox.Text;
+                secrets.sandboxServerBEID = sandboxBeidTxtBox.Text;
+                secrets.sandboxServerKey = sandboxKeyTxtBox.Text;
+                secrets.sandboxUserAddress = sandboxUserServerTxtBox.Text;
+                secrets.sandboxUserUsername = sandboxUserUsernameTxtBox.Text;
+                secrets.sandboxUserPwd = sandboxUserPwdTxtBox.Text;
+
+                secrets.prodServerAddress = prodServerTxtBox.Text;
+                secrets.prodServerBEID = prodBeidTxtBox.Text;
+                secrets.prodServerKey = prodKeyTxtBox.Text;
+                secrets.prodUserAddress = prodUserServerTxtBox.Text;
+                secrets.prodUserUsername = prodUserUsernameTxtBox.Text;
+                secrets.prodUserPwd = prodUserPwdTxtBox.Text;
+
+                var updatedSecretsJson = JsonConvert.SerializeObject(secrets, Formatting.Indented);
+                File.WriteAllText(secretsPath, updatedSecretsJson);
+            }
+
+            //prodAdminBtn.Checked = false;
+            //sandboxAdminBtn.Checked = false;
+            //prodUserBtn.Checked = false;
+            //sandboxUserBtn.Checked = true;
+
+            if (prodAdminBtn.Checked == true)
+                globalSettings.addressBookSelection = (int) Selection.prodServer;
+            else if (sandboxAdminBtn.Checked == true)
+                globalSettings.addressBookSelection = (int) Selection.sandboxServer;
+            else if (prodUserBtn.Checked == true)
+                globalSettings.addressBookSelection = (int) Selection.prodUser;
+            else
+                globalSettings.addressBookSelection = (int) Selection.sandboxUser;
+
             this.Close();
         }
     }
